+++ mapie/tests/test_estimators.py
@@ -6,14 +6,14 @@ List of tests:
 - Test created attributes depending on the method
 - Test output and results
 """
-from typing import Any
+from typing import Any, Union
 
 import pytest
 import numpy as np
 from sklearn.datasets import make_regression
 from sklearn.dummy import DummyRegressor
 from sklearn.linear_model import LinearRegression
-from sklearn.model_selection import LeaveOneOut, Kfold
+from sklearn.model_selection import LeaveOneOut, KFold
 from sklearn.exceptions import NotFittedError
 from sklearn.utils.estimator_checks import parametrize_with_checks
 
@@ -29,17 +29,25 @@ SEED = 59
 np.random.seed(SEED)
 y_reg = y_reg + np.random.normal(0, 1, 500)
 
-ALL_METHODS = ["naive", "base", "plus", "minmax"]
-NAIVE_METHOD = "naive"
-NON_NAIVE_METHODS = ["base", "plus", "minmax"]
+
+METHODS = ["naive", "base", "plus", "minmax"]
+STRATEGIES = {
+    "naive": dict(method="naive"),
+    "jackknife": dict(method="base", cv=LeaveOneOut()),
+    "jackknife_plus": dict(method="plus", cv=LeaveOneOut()),
+    "jackknife_minmax": dict(method="minmax", cv=LeaveOneOut()),
+    "cv": dict(method="base", cv=KFold(n_splits=3, shuffle=True, random_state=1)),
+    "cv_plus": dict(method="plus", cv=KFold(n_splits=3, shuffle=True, random_state=1)),
+    "cv_minmax": dict(method="minmax", cv=KFold(n_splits=3, shuffle=True, random_state=1)),
+}
 EXPECTED_WIDTHS = {
     "naive": 3.76,
     "jackknife": 3.85,
     "jackknife_plus": 3.86,
     "jackknife_minmax": 3.91,
-    "cv": 3.92,
-    "cv_plus": 3.99,
-    "cv_minmax": 4.13
+    "cv": 4.04,
+    "cv_plus": 4.07,
+    "cv_minmax": 4.30
 }
 EXPECTED_COVERAGES = {
     "naive": 0.952,
@@ -50,7 +58,6 @@ EXPECTED_COVERAGES = {
     "cv_plus": 0.956,
     "cv_minmax": 0.966
 }
-
 SKLEARN_EXCLUDED_CHECKS = {
     "check_regressors_train",
     "check_pipeline_consistency",
@@ -58,18 +65,19 @@ SKLEARN_EXCLUDED_CHECKS = {
 }
 
 
-# TODO:
-# beware of ALL_METHODS, you should also check for cv et return_pred
-# rethink all tests, especially the last ones with expected widths/coverage
-# assert checks are called during fit
-# assert checks are called during predict
-# assert all corner cases of predict
-# update documentation
-# update examples
-
 def test_initialized() -> None:
     """Test that initialization does not crash."""
     MapieRegressor()
+
+
+def test_default_parameters() -> None:
+    """Test default values of input parameters."""
+    mapie = MapieRegressor()
+    assert mapie.estimator is None
+    assert mapie.alpha == 0.1
+    assert mapie.method == "plus"
+    assert mapie.cv is None
+    assert not mapie.ensemble
 
 
 def test_fit() -> None:
@@ -92,17 +100,7 @@ def test_no_fit_predict() -> None:
         mapie.predict(X_toy)
 
 
-def test_default_parameters() -> None:
-    """Test default values of input parameters."""
-    mapie = MapieRegressor(estimator=DummyRegressor())
-    assert mapie.estimator is None
-    assert mapie.alpha == 0.1
-    assert mapie.method == "plus"
-    assert mapie.cv == 5
-    assert not mapie.ensemble
-
-
-@pytest.mark.parametrize("estimator", [0, "estimator", Kfold()])
+@pytest.mark.parametrize("estimator", [0, "estimator", KFold(), ["a", "b"]])
 def test_invalid_estimator(estimator: Any) -> None:
     """Test that invalid estimators raise errors."""
     mapie = MapieRegressor(estimator=estimator)
@@ -117,17 +115,17 @@ def test_none_estimator() -> None:
     assert isinstance(mapie.single_estimator_, LinearRegression)
 
 
-@pytest.mark.parametrize("method", ALL_METHODS)
-def test_valid_estimator(method: str) -> None:
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+def test_valid_estimator(strategy: str) -> None:
     """Test that valid estimators are not corrupted."""
-    mapie = MapieRegressor(estimator=DummyRegressor(), method=method)
+    mapie = MapieRegressor(estimator=DummyRegressor(), **STRATEGIES[strategy])
     mapie.fit(X_toy, y_toy)
     assert isinstance(mapie.single_estimator_, DummyRegressor)
     for estimator in mapie.estimators_:
         assert isinstance(estimator, DummyRegressor)
 
 
-@pytest.mark.parametrize("alpha", [-1, 0, 1, 2])
+@pytest.mark.parametrize("alpha", [-1, 0, 1, 2, 2.5, "a", ["a", "b"]])
 def test_invalid_alpha(alpha: int) -> None:
     """Test that invalid alphas raise errors."""
     mapie = MapieRegressor(alpha=alpha)
@@ -135,14 +133,14 @@ def test_invalid_alpha(alpha: int) -> None:
         mapie.fit(X_toy, y_toy)
 
 
-@pytest.mark.parametrize("alpha", np.linspace(0, 1, 5))
+@pytest.mark.parametrize("alpha", np.linspace(0.01, 0.99, 5))
 def test_valid_alpha(alpha: int) -> None:
     """Test that valid alphas raise no errors."""
     mapie = MapieRegressor(alpha=alpha)
     mapie.fit(X_toy, y_toy)
 
 
-@pytest.mark.parametrize("method", ["dummy", "cv_dummy", "jackknife_dummy", "dummy_plus", "dummy_minmax"])
+@pytest.mark.parametrize("method", [0, 1, "jackknife", "cv", ["base", "plus"]])
 def test_invalid_method(method: str) -> None:
     """Test that invalid methods raise errors."""
     mapie = MapieRegressor(method=method)
@@ -150,23 +148,23 @@ def test_invalid_method(method: str) -> None:
         mapie.fit(X_toy, y_toy)
 
 
-@pytest.mark.parametrize("method", ALL_METHODS)
+@pytest.mark.parametrize("method", METHODS)
 def test_valid_method(method: str) -> None:
     """Test that valid methods raise no errors."""
     mapie = MapieRegressor(method=method)
     mapie.fit(X_toy, y_toy)
 
 
-@pytest.mark.parametrize("ensemble", ["dummy", 1, 2., [1, 2]])
+@pytest.mark.parametrize("ensemble", ["dummy", 0, 1, 2.5, [1, 2]])
 def test_invalid_ensemble(ensemble: Any) -> None:
     """Test that invalid return_pred raise errors."""
-    mapie = MapieRegressor(DummyRegressor(), ensemble=ensemble)
+    mapie = MapieRegressor(ensemble=ensemble)
     with pytest.raises(ValueError, match=r".*Invalid ensemble.*"):
         mapie.fit(X_toy, y_toy)
 
 
 @pytest.mark.parametrize("ensemble", [True, False])
-def test_valid_ensemble(ensemble: str) -> None:
+def test_valid_ensemble(ensemble: bool) -> None:
     """Test that valid ensemble raise no errors."""
     mapie = MapieRegressor(ensemble=ensemble)
     mapie.fit(X_toy, y_toy)
@@ -180,71 +178,92 @@ def test_invalid_cv(cv: Any) -> None:
         mapie.fit(X_toy, y_toy)
 
 
-@pytest.mark.parametrize("cv", [None, 2, 3, 4, 5, 10, Kfold(), LeaveOneOut()])
+@pytest.mark.parametrize("cv", [None, 2, KFold(), LeaveOneOut()])
 def test_valid_cv(cv: Any) -> None:
     """Test that valid cv raise no errors."""
     mapie = MapieRegressor(cv=cv)
     mapie.fit(X_toy, y_toy)
 
 
-@pytest.mark.parametrize("method", ALL_METHODS)
+@parametrize_with_checks([MapieRegressor()])  # type: ignore
+def test_sklearn_compatible_estimator(estimator: Any, check: Any) -> None:
+    """Check compatibility with sklearn, using sklearn estimator checks API."""
+    if check.func.__name__ not in SKLEARN_EXCLUDED_CHECKS:
+        check(estimator)
+
+
+@pytest.mark.parametrize("method", METHODS)
 def test_fit_attributes(method: str) -> None:
     """Test fit attributes shared by all PI methods."""
     mapie = MapieRegressor(method=method)
     mapie.fit(X_toy, y_toy)
+    assert hasattr(mapie, 'n_features_in_')
     assert hasattr(mapie, 'single_estimator_')
     assert hasattr(mapie, 'estimators_')
     assert hasattr(mapie, 'residuals_')
     assert hasattr(mapie, 'k_')
 
 
-@pytest.mark.parametrize("method", ALL_METHODS)
-def test_predict_output_shape(method: str) -> None:
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+@pytest.mark.parametrize("ensemble", [True, False])
+def test_predict_output_shape(strategy: str, ensemble: bool) -> None:
     """Test predict output shape."""
-    mapie = MapieRegressor(method=method)
+    mapie = MapieRegressor(ensemble=ensemble, **STRATEGIES[strategy])
     mapie.fit(X_reg, y_reg)
-    assert mapie.predict(X_reg).shape[0] == X_reg.shape[0]
-    assert mapie.predict(X_reg).shape[1] == 3
+    y_preds = mapie.predict(X_reg)
+    assert y_preds.shape[0] == X_reg.shape[0]
+    assert y_preds.shape[1] == 3
 
 
-@pytest.mark.parametrize("method", ALL_METHODS)
-def test_linear_confidence_interval(method: str) -> None:
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+@pytest.mark.parametrize("ensemble", [True, False])
+def test_prediction_between_low_up(strategy: str, ensemble: bool) -> None:
+    """Test that prediction lies between low and up prediction intervals."""
+    mapie = MapieRegressor(ensemble=ensemble, **STRATEGIES[strategy])
+    mapie.fit(X_reg, y_reg)
+    y_preds = mapie.predict(X_reg)
+    y_pred, y_pred_low, y_pred_up = y_preds.T
+    assert (y_pred >= y_pred_low).all()
+    assert (y_pred <= y_pred_up).all()
+
+
+@pytest.mark.parametrize("method", ["plus", "minmax"])
+@pytest.mark.parametrize("cv", [LeaveOneOut(), KFold(n_splits=3, shuffle=True, random_state=1)])
+def test_prediction_ensemble(method: str, cv: Union[LeaveOneOut, KFold]) -> None:
+    """Test that predictions differs when ensemble is True/False, but not prediction intervals."""
+    mapie = MapieRegressor(method=method, cv=cv, ensemble=True)
+    mapie.fit(X_reg, y_reg)
+    y_preds_ensemble = mapie.predict(X_reg)
+    mapie.ensemble = False
+    y_preds_no_ensemble = mapie.predict(X_reg)
+    np.testing.assert_almost_equal(y_preds_ensemble[:, 1], y_preds_no_ensemble[:, 1])
+    np.testing.assert_almost_equal(y_preds_ensemble[:, 2], y_preds_no_ensemble[:, 2])
+    with pytest.raises(AssertionError):
+        np.testing.assert_almost_equal(y_preds_ensemble[:, 0], y_preds_no_ensemble[:, 0])
+
+
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+@pytest.mark.parametrize("ensemble", [True, False])
+def test_linear_data_confidence_interval(strategy: str, ensemble: bool) -> None:
     """
     Test that MapieRegressor applied on a linear regression model
     fitted on a linear curve results in null uncertainty.
     """
-    mapie = MapieRegressor(estimator=LinearRegression(), method=method, n_splits=3)
+    mapie = MapieRegressor(ensemble=ensemble, **STRATEGIES[strategy])
     mapie.fit(X_toy, y_toy)
     y_preds = mapie.predict(X_toy)
-    y_low, y_up = y_preds[:, 1], y_preds[:, 2]
-    np.testing.assert_almost_equal(y_up, y_low)
+    _, y_pred_low, y_pred_up = y_preds.T
+    np.testing.assert_almost_equal(y_pred_up, y_pred_low)
 
 
-@pytest.mark.parametrize("ensemble", [True, False])
-def test_prediction_between_low_up(ensemble: bool) -> None:
-    """Test that prediction lies between low and up prediction intervals."""
-    mapie = MapieRegressor(estimator=LinearRegression(), ensemble=ensemble)
-    mapie.fit(X_reg, y_reg)
-    y_preds = mapie.predict(X_reg)
-    y_pred, y_low, y_up = y_preds[:, 0], y_preds[:, 1], y_preds[:, 2]
-    assert (y_pred >= y_low).all() & (y_pred <= y_up).all()
-
-
-@pytest.mark.parametrize("method", ALL_METHODS)
-def test_linreg_results(method: str) -> None:
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+def test_linear_regression_results(strategy: str) -> None:
     """Test expected PIs for a multivariate linear regression problem with fixed random seed."""
-    mapie = MapieRegressor(estimator=LinearRegression(), method=method, alpha=0.05, random_state=SEED)
+    mapie = MapieRegressor(estimator=LinearRegression(), alpha=0.05, **STRATEGIES[strategy])
     mapie.fit(X_reg, y_reg)
     y_preds = mapie.predict(X_reg)
-    preds_low, preds_up = y_preds[:, 1], y_preds[:, 2]
-    width_mean = (preds_up-preds_low).mean()
-    coverage = coverage_score(y_reg, preds_low, preds_up)
-    np.testing.assert_almost_equal(width_mean, EXPECTED_WIDTHS[method], 2)
-    np.testing.assert_almost_equal(coverage, EXPECTED_COVERAGES[method], 2)
-
-
-@parametrize_with_checks([MapieRegressor(LinearRegression())])  # type: ignore
-def test_sklearn_compatible_estimator(estimator: Any, check: Any) -> None:
-    """Check compatibility with sklearn, using sklearn estimator checks API."""
-    if check.func.__name__ not in SKLEARN_EXCLUDED_CHECKS:
-        check(estimator)
+    _, y_pred_low, y_pred_up = y_preds.T
+    width_mean = (y_pred_up - y_pred_low).mean()
+    coverage = coverage_score(y_reg, y_pred_low, y_pred_up)
+    np.testing.assert_almost_equal(width_mean, EXPECTED_WIDTHS[strategy], 2)
+    np.testing.assert_almost_equal(coverage, EXPECTED_COVERAGES[strategy], 2)